refactor: restructure language and sidebar components for improved organization
