+++ src/composables/useSidebarController.ts
