+++ src/layouts/TheHeader.vue
@@ -1,16 +1,18 @@
 <script setup lang="ts">
-import { useSidebarController } from '@/features/composables/useSidebarController'
-import ChangeLang from '@/features/components/ChangeLang.vue'
+import { useSidebarController } from '@/composables/useSidebarController'
+import ChangeLang from '@/components/ChangeLang.vue'
 
 const { toggleSidebar } = useSidebarController()
 </script>
 <template>
   <v-app-bar color="background" density="compact" flat class="d-flex align-center">
-    <template v-slot:prepend>
+    <template #prepend>
       <v-btn class="text-primary" icon @click="toggleSidebar">
         <v-icon>mdi-menu</v-icon>
       </v-btn>
     </template>
-    <ChangeLang />
+    <template #append>
+      <ChangeLang />
+    </template>
   </v-app-bar>
 </template>
